+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2018-03-20 Improved SQL statement for function block_cohortspecifichtml_cohorts_is_member.
 * 2018-03-20 Minor improvements to language file and behat tests.
 
 ### Release v3.2-r1 - v3.2-r3